GameType refactor part 3
Now direction of a Move has a different meaning depending on the value of the
GameType enum. If a Move is Forward, then the move is a push in that direction.
If the Move is Reverse, then the move is a pull in that direction. Pull in a
direction now actually means pulling in the given direction (a pull to Left
actually means moving the box left and placing the player on the left of the
destination of the pulled box).